+++ include/librealsense2/hpp/rs_safety_types.hpp
@@ -1,0 +1,9 @@
+// License: Apache 2.0. See LICENSE file in root directory.
+// Copyright(c) 2022 Intel Corporation. All Rights Reserved.
+
+#ifndef LIBREALSENSE_RS2_SAFETY_TYPES_HPP
+#define LIBREALSENSE_RS2_SAFETY_TYPES_HPP
+
+#include "../h/rs_safety_types.h"
+
+#endif // LIBREALSENSE_RS2_SAFETY_TYPES_HPP
