+++ src/ds/ds-private.h
@@ -36,9 +36,6 @@ namespace librealsense
         const uint8_t DS5_LED_PWR = 0xE;
         const uint8_t DS5_THERMAL_COMPENSATION = 0xF;
         
-        //DS6 safety XU identifiers
-        const uint16_t DS6_ACTIVE_SAFETY_PRESET = 0x02;
-
         enum class ds_device_type
         {
             ds5,
