+++ src/media/playback/playback_device.cpp
@@ -763,6 +763,7 @@ bool playback_device::try_extend_snapshot(std::shared_ptr<extension_snapshot>& e
     case RS2_EXTENSION_COLOR_SENSOR: return try_extend<color_sensor>(e, ext);
     case RS2_EXTENSION_MOTION_SENSOR: return try_extend<motion_sensor>(e, ext);
     case RS2_EXTENSION_FISHEYE_SENSOR: return try_extend<fisheye_sensor>(e, ext);
+    case RS2_EXTENSION_SAFETY_SENSOR: return try_extend<safety_sensor>(e, ext);
     case RS2_EXTENSION_UNKNOWN: //[[fallthrough]]
     case RS2_EXTENSION_COUNT:   //[[fallthrough]]
     default:
