+++ src/sensor.h
@@ -197,11 +197,8 @@ namespace librealsense
 
         void create_snapshot(std::shared_ptr<safety_sensor>& snapshot) const override;
         void enable_recording(std::function<void(const safety_sensor&)> recording_function) override {};
-        virtual void set_safety_preset(int index, const rs2_safety_preset& sp) const {}
-        virtual rs2_safety_preset get_safety_preset(int index) const {
-            rs2_safety_preset r;
-            return r;
-        }
+        virtual void set_safety_preset(int index, const rs2_safety_preset& sp) const = 0;
+        virtual rs2_safety_preset get_safety_preset(int index) const = 0;
     };
     MAP_EXTENSION(RS2_EXTENSION_SAFETY_SENSOR, librealsense::safety_sensor);
 
@@ -224,7 +221,7 @@ namespace librealsense
             //empty
         }
 
-        void set_safety_preset(int index, const rs2_safety_preset& sp)  {};
+        void set_safety_preset(int index, const rs2_safety_preset& sp)  const {};
         rs2_safety_preset get_safety_preset(int index) const
         {
             rs2_safety_preset r;
