+++ unit-tests/live/safety/test-safety-preset-get-set-xu-cmd.py
@@ -1,0 +1,63 @@
+# License: Apache 2.0. See LICENSE file in root directory.
+# Copyright(c) 2022 Intel Corporation. All Rights Reserved.
+
+#test:donotrun
+#test:device S585
+
+import pyrealsense2 as rs
+from rspy import test
+
+#############################################################################################
+# Tests
+#############################################################################################
+
+test.start("Init")
+try:
+    ctx = rs.context()
+    dev = ctx.query_devices()[0]
+    safety_sensor = dev.first_safety_sensor()
+except:
+    test.unexpected_exception()
+test.finish()
+
+#############################################################################################
+
+test.start("Check if safety sensor supports the option")
+test.check_equal(safety_sensor.supports(rs.option.safety_preset_active_index), True)
+test.finish()
+
+#############################################################################################
+test.start("Valid get/set scenario")
+
+# default index at start should be 0
+current_index = safety_sensor.get_option(rs.option.safety_preset_active_index)
+test.check_equal(current_index, 0)
+
+safety_sensor.set_option(rs.option.safety_preset_active_index, 1)
+current_index = safety_sensor.get_option(rs.option.safety_preset_active_index)
+test.check_equal(current_index, 1)
+
+safety_sensor.set_option(rs.option.safety_preset_active_index, 20)
+current_index = safety_sensor.get_option(rs.option.safety_preset_active_index)
+test.check_equal(current_index, 20)
+
+current_index = safety_sensor.get_option(rs.option.safety_preset_active_index)
+safety_sensor.set_option(rs.option.safety_preset_active_index, 63)
+test.check_equal(current_index, 63)
+
+test.finish()
+
+#############################################################################################
+
+test.start("Invalid set - index out of range")
+try:
+    safety_sensor.set_option(rs.option.safety_preset_active_index, 64)
+except:
+    pass
+else:
+    test.unexpected_exception()
+test.finish()
+
+#############################################################################################
+
+test.print_results_and_exit()
