+++ wrappers/nodejs/src/addon.cpp
@@ -4865,7 +4865,7 @@ void InitModule(v8::Local<v8::Object> exports) {
   _FORCE_SET_ENUM(RS2_OPTION_RECEIVER_SENSITIVITY);
   _FORCE_SET_ENUM(RS2_OPTION_AUTO_EXPOSURE_LIMIT_TOGGLE);
   _FORCE_SET_ENUM(RS2_OPTION_AUTO_GAIN_LIMIT_TOGGLE);
-  _FORCE_SET_ENUM(RS2_OPTION_ACTIVE_SAFETY_PRESET_INDEX);
+  _FORCE_SET_ENUM(RS2_OPTION_SAFETY_PRESET_ACTIVE_INDEX);
   _FORCE_SET_ENUM(RS2_OPTION_COUNT);
 
   // rs2_camera_info
